Anpassungen SupermarketGEO-Scheduler

+++ src/main/java/de/berlin/htw/usws/metamodel/SupermarketGEO_.java
@@ -14,12 +14,12 @@ public abstract class SupermarketGEO_ extends BaseEntity_ {
 	public static volatile SingularAttribute<SupermarketGEO, Supermarket> supermarket;
 	public static volatile SingularAttribute<SupermarketGEO, String> zip;
 	public static volatile SingularAttribute<SupermarketGEO, String> supermarketName;
-	public static volatile SingularAttribute<SupermarketGEO, Float> lng;
+	public static volatile SingularAttribute<SupermarketGEO, Double> lng;
 	public static volatile SingularAttribute<SupermarketGEO, String> housenumber;
 	public static volatile SingularAttribute<SupermarketGEO, String> city;
 	public static volatile SingularAttribute<SupermarketGEO, String> street;
 	public static volatile SingularAttribute<SupermarketGEO, String> phonenumber;
-	public static volatile SingularAttribute<SupermarketGEO, Float> let;
+	public static volatile SingularAttribute<SupermarketGEO, Double> let;
 	public static volatile SingularAttribute<SupermarketGEO, String> marketID;
 
 }
